Plots global importance (i.e. cumulative importance of all features for all response variables) and individual response importance.

Plots global importance (i.e. cumulative importance of all features for all response variables) and individual response importance.

interpret_Mrvi(
  VI,
  modelPerf,
  Y,
  X,
  groupCov = NULL,
  cutoff = 0.2,
  mode = "regression"
)

Arguments

VI

A dataframe data frame generated from mrvip function

modelPerf

A dataframe data frame generated by modelPerf function

Y

A dataframe response data set

X

A dataframe feature data set

cutoff

A numeric determines Mathews correlation coefficient (mcc) threshold for displaying individual reponse model. Default is 0.5

mode

character'classification' or 'regression' i.e., is the generative model a regression or classification?

Details

1st plot: and individual response feature plots (2nd plot). Requires object generated from the R function mrvip as well as feature data (Y). Variables also need to be grouped to allow for easier interpretation.