Compare and benchmark disease outbreak risk among and within groups

Compare and benchmark disease outbreak risk among and within groups

mrBenchmark(
  data = "data",
  Y = "class",
  pred = "predicted",
  group = "group1",
  label_by = "ID",
  type = "internal"
)

Arguments

data

A character object name of data frame

Y

A character column name of variable containing outcome

pred

A character column name of variable containing model predicted values

group

A character column name of variable that individuals should be grouped by

label_by

A character column name of variable representing the individual units. If stated, these will be labeled on the ggplot. By default labels will not be included

type

A character specify within group "internal" or among group "external" benchmarking

Examples

mB <- mrBenchmark(data=data, Y='class')
#> Error in parse(text = data): cannot coerce type 'closure' to vector of type 'character'