Wrapper to calculate performance metrics (Mathews correlation coefficient, sensitivity and specificity) for each model for each response variable.

Wrapper to calculate performance metrics (Mathews correlation coefficient, sensitivity and specificity) for each model for each response variable.

Usage

mrIMLperformance(yhats, Model, Y, mode = "regression")

Arguments

yhats

A list is the list generated by mrIMLpredicts

Model

A list the model used to generate the yhats object

Y

A dataframe is a response variable data set (species, SNPs etc).

mode

character'classification' or 'regression' i.e., is the generative model a regression or classification?

Details

Outputs a dataframe of commonly used metric that can be used to compare model performance of classification models. Performance metrics are based on testing data. But MCC is useful (higher numbers = better fit)

Examples

if (FALSE) { # \dontrun{
ModelPerf <- mrIMLperformance(yhats, Model=model1, Y=Y) } # }