Bootstrap Partial Dependence plots

This function bootstraps model predictions and generates partial dependence plots for each response variable. It also creates a combined plot for the top variables of interest.

mrPD_bootstrap(mrBootstrap_obj, vi_obj, X, Y, target, global_top_var = 2)

Arguments

mrBootstrap_obj

A list of model bootstraps generated using mrBootstrap function.

vi_obj

Variable Importance data.

X

The predictor data.

Y

The response data.

target

The target variable for generating plots.

global_top_var

The number of top variables to consider (default: 2).

Value

A list containing the partial dependence plots for each response variable and a combined plot.

Examples