Plot Model Performance Comparison

Create visualizations to compare the performance of two models based on their performance metrics generated by mrIMLperformance.

Usage

mrPerformancePlot(
  ModelPerf1 = NULL,
  ModelPerf2 = NULL,
  mode = "classification"
)

Arguments

ModelPerf1, ModelPerf2

Two data frames of model performance metrics to compare. The data frames are created by mrIMLperformance, see Examples.

mode

A character string describing the mode of the models. Should be either "regression" or "classification". The default is "classification".

Value

A list containing:

$performance_plot: A box plot of model performance metrics.

$performance_diff_plot: A bar plot of the differences in performance metrics.

$performance_diff_df: A data frame in wide format containing model performance metrics and their differences.

Examples

MR_perf_rf <-  mrIML::mrIML_bird_parasites_RF %>%
  mrIMLperformance()
MR_perf_lm <- mrIML::mrIML_bird_parasites_LM%>%
  mrIMLperformance()

perf_comp <- mrPerformancePlot(
  ModelPerf1 = MR_perf_rf,
  ModelPerf2 = MR_perf_lm
)