Plots global interactions as well as individual response interaction importance.

Plots global interactions as well as individual response interaction importance.

Usage

mrPlot_interactions(interactions, X, Y, top_ranking = 3, top_response = 10)

Arguments

interactions

A dataframe data frame generated from mrInteractioms function

X

A dataframe feature data set

Y

A dataframe response data set

top_ranking

A numeric determines how many of the strongest feature interacions to view/include

top_response

A numeric how many of the response variables with the strongest interactions to view

Details

1st plot: Barplots showing the mean and cumulative importance of each of the top pairs of interactions in the model. 2nd plot: Barplot of the responses with the strongest interactions 3rd plot: Barplots of the strongest interactions for each of the top response variables. interactions <-mrInteractions(yhats, X, Y) #this is computationally intensive so multicores are needed. If stopped prematurely - have to reload things mrPlot_interactions(Interact, X,Y, top_ranking = 3, top_response=3)