Calculates resistance components from a list of pairwise resistance surfaces.

Calculates resistance components from a list of pairwise resistance surfaces.

Usage

resist_components(foldername = foldername, p_val = p_val)

Arguments

foldername

A character this is the location where the resistance surfaces are stored.

p_val

A numeric this sets the significance threshold for axes in explaining variance in the original resistance matrix based on redundancy analysis. In effect this filters out axes that dont explain variance.

Details

Outputs a data frame of significant resistance components for each matrix in the target folder. These data can be combined with non-pairwise matrix data.

Examples

Y <- resist_components(filename = 'Bobcat_cs_matrices', p_val=0.01)
#> Error in resist_components(filename = "Bobcat_cs_matrices", p_val = 0.01): unused argument (filename = "Bobcat_cs_matrices")