mrProfileplot: Wrapper to plot mutlti-response model agnostic profile plots (partial dependences and accumulated local effects).

mrProfileplot: Wrapper to plot mutlti-response model agnostic profile plots (partial dependences and accumulated local effects).

Usage

mrProfileplot(profileData, sdthresh = 0.05)

Arguments

profileData

dataframe is the data generated from the flashlight packages 'light_profile' function.

'sdthresh'

numeric value used to filter responses that are not changing across the values of the feature (based on standard deviation)

Details

The aim of this function is to plot (1) a reduced set of response variables that are responding to the feature of choice (plot 1) and the average ALE of partial dependency for all responses combined (plot 2). When there are many responses plot 1 makes interpretation easier by focussing on the the responses changing the most with a feature. The feature selected and plot type must be specified 'light_profile' function

Examples

flashlightObj <- mrFlashlight(yhats, X, Y, response = "multi")
#> Error in eval(expr, envir, enclos): object 'yhats' not found

profileData_pd <- light_profile(flashlightObj, v = "Grassland") #partial dependencies
#> Error in light_profile(flashlightObj, v = "Grassland"): could not find function "light_profile"
profileData_ale <- light_profile(flashlightObj, v = "Grassland", type = "ale") #acumulated local effects
#> Error in light_profile(flashlightObj, v = "Grassland", type = "ale"): could not find function "light_profile"
mrProfileplot(profileData_pd , sdthresh =0.05)
#> Error in eval(expr, envir, enclos): object 'profileData_pd' not found
mrProfileplot(profileData_ale , sdthresh =0.05)
#> Error in eval(expr, envir, enclos): object 'profileData_ale' not found