Calculates and helps interpret variable importance for mrIML models.

Calculates and helps interpret variable importance for mrIML models.

Usage

mrvip(
  yhats = NULL,
  mrBootstrap_obj = NULL,
  X = X,
  X1 = NULL,
  Y = Y,
  mode = "classification",
  threshold = 0.1,
  global_top_var = 10,
  local_top_var = 5,
  taxa = NULL,
  ModelPerf = ModelPerf,
  plot.pca = T
)

Arguments

yhats

A list of model predictions.

mrBootstrap_obj

The object containing model bootstrapping results.

X

The predictor data.

X1

A dataframe extra predictor set used in each model. For the MrIML Joint species distribution model (JSDM) this is just a copy of the response data.

Y

The response data.

mode

character'classification' or 'regression' i.e., is the generative model a regression or classification?

threshold

The threshold for model performance (AUC) below which variables are filtered (default: 0.1).

global_top_var

The number of top global variables to display (default: 10).

local_top_var

The number of top local variables for each response to display (default: 5).

ModelPerf

A list containing model performance metrics for each response variable.

Value

A list containing variable importance data and a combined plot.

Examples